# apispec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cashflow Restful API",
    "description": "API Spec for Cashflow Restful API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://rumah-dermawan-backend-6d8a2add9817.herokuapp.com/api/v1"
    }
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "user auth"
        ],
        "description": "register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/register request body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "register success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Account has been register"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateUpdateUserResponse"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "input not valid, failed for register",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 405
                    },
                    "message": {
                      "type": "string",
                      "example": "input not valid !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "when fail to post to db and generate token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "register failed, please try again !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user auth"
        ],
        "description": "login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login request body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Login success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateUpdateUserResponse"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "input not valid, failed for login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 405
                    },
                    "message": {
                      "type": "string",
                      "example": "input not valid !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "error to login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "login failed, please try again !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/period/create": {
      "post": {
        "tags": [
          "period"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "insert period",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "required": true,
                    "type": "string",
                    "example": "2023"
                  },
                  "month": {
                    "required": true,
                    "type": "string",
                    "example": "january"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succes create period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": "200"
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "year": {
                          "type": "string",
                          "example": "2023"
                        },
                        "month": {
                          "type": "string",
                          "example": "january"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "input not valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 405
                    },
                    "message": {
                      "type": "string",
                      "example": "input not valid !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "error to create period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "create failed, please try again !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/period/{periodId}": {
      "put": {
        "tags": [
          "period"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "update period",
        "description": "update period",
        "parameters": [
          {
            "in": "path",
            "name": "periodId",
            "description": "period id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "required": true,
                    "type": "string",
                    "example": "2023"
                  },
                  "month": {
                    "required": true,
                    "type": "string",
                    "example": "january"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "succes create period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": "200"
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/object period"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "input not valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 405
                    },
                    "message": {
                      "type": "string",
                      "example": "input not valid !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "error to update period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "create failed, please try again !"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "period"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "delete period",
        "description": "update period",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string",
              "format": "token"
            },
            "required": true,
            "description": "token from login"
          },
          {
            "in": "path",
            "name": "periodId",
            "description": "period id"
          }
        ],
        "responses": {
          "200": {
            "description": "success delete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": "200"
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateUpdateUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "email": {
            "type": "string",
            "example": "dzakyhdr@gmail.com"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}